+++ app/src/main/java/com/kionavani/todotask/di/appComponent/AppComponent.kt
@@ -1,10 +1,9 @@
 package com.kionavani.todotask.di.appComponent
 
 import android.content.Context
-import com.kionavani.todotask.data.remote.DataFetchWorker
+import com.kionavani.todotask.ui.DataFetchWorker
 import com.kionavani.todotask.di.AppScope
 import com.kionavani.todotask.di.screensComponent.ScreenComponent
-import com.kionavani.todotask.di.screensComponent.ViewModelModule
 import dagger.BindsInstance
 import dagger.Component
 import kotlinx.coroutines.CoroutineScope
