+++ app/src/main/java/com/kionavani/todotask/ui/DataFetchWorker.kt
@@ -1,25 +1,27 @@
-package com.kionavani.todotask.data.remote
+package com.kionavani.todotask.ui
 
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.kionavani.todotask.data.NetworkMonitor
 import com.kionavani.todotask.domain.TodoItemsRepository
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
 
 /**
  * Класс Worker'a, отвечающий за синхронизацию данных в фоне
  */
-class DataFetchWorker @AssistedInject constructor(
-    @Assisted context: Context,
-    @Assisted workerParams: WorkerParameters,
-    private val repository: TodoItemsRepository,
-    private val networkMonitor: NetworkMonitor
+class DataFetchWorker (
+    context: Context,
+    workerParams: WorkerParameters
 ) : CoroutineWorker(context, workerParams) {
+    @Inject
+    lateinit var repository: TodoItemsRepository
+    @Inject
+    lateinit var networkMonitor: NetworkMonitor
 
     override suspend fun doWork(): Result = withContext(Dispatchers.IO) {
         try {
