+++ app/src/main/java/com/kionavani/todotask/ui/NetworkMonitor.kt
@@ -1,4 +1,4 @@
-package com.kionavani.todotask.data
+package com.kionavani.todotask.ui
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -6,7 +6,6 @@ import android.net.Network
 import android.net.NetworkCapabilities
 import android.net.NetworkRequest
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 
 /**
