+++ app/src/main/java/com/kionavani/todotask/ui/composable/Preview.kt
@@ -7,7 +7,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
-import com.kionavani.todotask.data.NetworkMonitor
+import com.kionavani.todotask.ui.NetworkMonitor
 import com.kionavani.todotask.domain.ToDoItem
 import com.kionavani.todotask.domain.TodoItemsRepository
 import com.kionavani.todotask.ui.ResourcesProvider
